code: add i2shell.ino
This is a tiny terminal emulator for attiny85 that parses commands and
turns them to I2C read/write operations. It's tries to retain some
settings like the device's address and to make it as easy as possible
to send/receive data bytes from a shell using only the "socat" utility.

E.g.

  - initialise DS1307:
    echo "S68W0 0P" | socat - /dev/ttyACM1,rawer

  - read time on DS1307:
    echo "S 68 W 0 R 7" | socat - /dev/ttyACM1,rawer
    35 05 00 01 01 01 00

The supported commands are:
  - S <addr>  : start / resync to idle for address <addr>
  - W <byte>* : send byte(s) to addr above
  - P         : push (end of write sequence)
  - R <num>   : read <num> bytes

The DigiCDC virtual terminal is limited to 32 bytes and the TinyWireM
is also limited to a few tens of bytes so it's not possible to exchange
large data chunks at once, but overall it works reasonably well to even
dump registers for example:

  $ for i in {0..3}; do echo -n "${i}0: ";echo "w${i}0r10" | socat - /dev/ttyACM1,rawer,readbytes=50 ;done
  00: 27 54 00 01 01 01 00 03 B1 02 20 08 12 11 03 09
  10: 30 29 24 00 24 82 D4 14 00 41 A8 20 01 D4 20 18
  20: 00 C2 0A 00 B4 00 00 80 40 11 50 01 80 88 06 02
  30: 02 01 40 81 00 00 00 20 08 40 00 40 04 00 50 00

(readbytes=50 makes socat quit after receiving 50 chars, which saves it
from having to wait for the timeout after 16 hex bytes are dumped). The
whole dump above takes 1.33s.

It's also convenient to send sequences as multiple commands, such as
the sequence below that configures an Si5351A as an 1 MHz oscillator:

  socat - /dev/ttyACM1  <<< "S 60 W 3 ff P"
  socat - /dev/ttyACM1  <<< "W 10 80 80 80 80 80 80 80 80 P"
  socat - /dev/ttyACM1  <<< "W F 0 P"
  socat - /dev/ttyACM1  <<< "W 18 0 P"
  socat - /dev/ttyACM1  <<< "W 2 F0 P"
  socat - /dev/ttyACM1  <<< "W 1A FF FF 00 0A 00 F0 00 00 P"
  socat - /dev/ttyACM1  <<< "W 2A 00 01 01 2A 00 00 00 00 P"
  socat - /dev/ttyACM1  <<< "W B1 AC P"
  # Set reference load C: 6 pF = 0x12, 8 pF = 0x92, 10 pF = 0xD2
  socat - /dev/ttyACM1  <<< "W B7 D2 P"
  socat - /dev/ttyACM1  <<< "W 10 4F P"
  socat - /dev/ttyACM1  <<< "W 3 FE P"