code: improve interactivity and make the code more reliable
There's a limit in DigiCDC by which if we don't write often enough,
input characters are lost and may even hang. We had a workaround for
this but it results in emitting a bit too many spaces.

Here instead we only emit a char when we read nothing, and if we have
not already emitted something since the last char was read. The char
sent is a CR if the last char printed was a CR or an LF, so that it
doesn't needlessly inject spaces at the beginning of lines (which is
annoying on the first line of dumps), and a space in other cases
(the only one that will not damage dumps nor confuse interactive
sessions).

A help command was added with character '?'. Finally the parser's
state could be simplified to start in ST_CMD.